* Lil fix for my last commit. Sorry :)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@13696 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/battle.c
@@ -2758,7 +2758,7 @@ struct Damage battle_calc_misc_attack(struct block_list *src,struct block_list *
 			cardfix=cardfix*(100-tsd->long_attack_def_rate)/100;
 
 		if (cardfix != 10000)
-			md.damage=md.damage*(cardfix/10000);
+			md.damage=md.damage*((double)cardfix/10000);
 	}
 
 	if (sd && (i = pc_skillatk_bonus(sd, skill_num)))
